вынесения функций для склонения слов в отдельную папку
