+++ clubs/lib/src/domain/entities/club/create_and_edit/club/condition_string.dart
@@ -51,29 +51,3 @@ class ConditionString extends StatelessWidget {
     );
   }
 }
-
-String inspectorWord(int count) {
-  int mod10 = count % 10;
-  int mod100 = count % 100;
-  
-  if (mod10 == 1 && mod100 != 11) {
-    return 'инспектор';
-  } else if (mod10 >= 2 && mod10 <= 4 && !(mod100 >= 12 && mod100 <= 14)) {
-    return 'инспектора';
-  } else {
-    return 'инспекторов';
-  }
-}
-
-String questionWord(int count) {
-  int mod10 = count % 10;
-  int mod100 = count % 100;
-  
-  if (mod10 == 1 && mod100 != 11) {
-    return 'вопрос';
-  } else if (mod10 >= 2 && mod10 <= 4 && !(mod100 >= 12 && mod100 <= 14)) {
-    return 'вопроса';
-  } else {
-    return 'вопросов';
-  }
-}
